# docs/swagger.yaml
basePath: /
definitions:
...
  models.Event:
    properties:
      event_name:
        type: string
      id:
        type: integer
    type: object
  models.EventCategory:
    properties:
      event_type:
        $ref: '#/definitions/models.EventType'
      event_type_id:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
  models.EventDetails:
    properties:
      address:
        type: string
      beneficiary_child:
        type: integer
      beneficiary_men:
        type: integer
      beneficiary_women:
        type: integer
      city:
        type: string
      country:
        type: string
      created_by:
        type: string
      created_on:
        type: string
      daily_end_time:
        type: string
      daily_start_time:
        type: string
      district:
        type: string
      end_date:
        type: string
      event_category:
        $ref: '#/definitions/models.EventCategory'
      event_category_id:
        type: integer
      event_type:
        $ref: '#/definitions/models.EventType'
      event_type_id:
        type: integer
      id:
        type: integer
      initiation_child:
        type: integer
      initiation_men:
        type: integer
      initiation_women:
        type: integer
      pincode:
        type: string
      post_office:
        type: string
      scale:
        type: string
      spiritual_orator:
        type: string
      start_date:
        type: string
      state:
        type: string
      theme:
        type: string
      updated_by:
        type: string
      updated_on:
        type: string
    type: object
  models.EventMedia:
    properties:
      company_email:
        type: string
      company_name:
        type: string
      company_website:
        type: string
      contact:
        type: string
      created_by:
        description: only set on create
        type: string
      created_on:
        type: string
      designation:
        type: string
      email:
        type: string
      event:
        $ref: '#/definitions/models.Event'
      event_id:
        type: integer
      first_name:
        type: string
      gender:
        type: string
      id:
        type: integer
      last_name:
        type: string
      media_coverage_type:
        $ref: '#/definitions/models.MediaCoverageType'
      media_coverage_type_id:
        type: integer
      middle_name:
        type: string
      prefix:
        type: string
      updated_by:
        type: string
      updated_on:
        type: string
    type: object
  models.EventType:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.MediaCoverageType:
    properties:
      id:
        type: integer
      media_type:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: vaish@gmail.com
    name: API Support
  description: API documentation for the DJJS Event Reporting Backend.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: DJJS Event Reporting API
  version: "1.0"
paths:
...
  /api/event-media:
    get:
      description: Retrieve all EventMedia records
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.EventMedia'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all Event Media
      tags:
      - EventMedia
    post:
      consumes:
      - application/json
      description: Create a new record in EventMedia table
      parameters:
      - description: Event Media Details
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.EventMedia'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.EventMedia'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create new Event Media
      tags:
      - EventMedia
  /api/event-media/{id}:
    delete:
      description: Delete a record by ID from EventMedia
      parameters:
      - description: Event Media ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete Event Media
      tags:
      - EventMedia
    put:
      consumes:
      - application/json
      description: Update an EventMedia record by ID
      parameters:
      - description: Event Media ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated details
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.EventMedia'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EventMedia'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update Event Media
      tags:
      - EventMedia
  /api/event-media/event/{event_id}:
    get:
      description: Get all Event Media records for a specific Event ID
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.EventMedia'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Event Media by Event ID
      tags:
      - EventMedia
  /api/events:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.EventDetails'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all events
      tags:
      - Events
    post:
      consumes:
      - application/json
      parameters:
      - description: Event payload
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/models.EventDetails'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new event
      tags:
      - Events
...
  /api/events/{id}:
    delete:
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete an event
      tags:
      - Events
    put:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated fields
        in: body
        name: event
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update an event
      tags:
      - Events
  /api/events/search:
    get:
      description: Search events by keyword
      parameters:
      - description: Search keyword
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.EventDetails'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Search events
      tags:
      - Events
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"